Move some tests to /sync/git/
