+++ shared/src/commonTest/kotlin/me/saket/press/shared/sync/git/FileNameRegisterTest.kt
@@ -1,4 +1,4 @@
-package me.saket.press.shared.sync
+package me.saket.press.shared.sync.git
 
 import assertk.assertThat
 import assertk.assertions.endsWith
